HTML/CSS for Navbar and Profile bar done
Need to implement requests to fetch the profile picture from the server depending on which user it is. Username, bio and post data already does this.

Also need to work out how to output posts to a grid pattern.

Two Modals are currently in use and can be accessed but do not achieve anything.